fix date handling in sheets plugin
I noticed an issue with handling of dates in CSV files in the sheets
plugin. If the `cellDates` option was set, and a date was provided
(with no timezone information) the date was off by one under daylight
savings (may be because I'm in GMT which is close to UTC). The internal
representation seems to be UTC, and setting `UTC: true` on the
`sheet_to_json` function fixes the issue. I don't think this should
cause issues with other formats.